ajustes no codigo

+++ src/main/java/com/api/stock/controller/PedidoController.java
@@ -41,7 +41,7 @@ public class PedidoController {
     }
 
     @GetMapping("/get/{id}")
-    public ResponseEntity<Object> getOnePedido(@PathVariable Integer id) {
+    public ResponseEntity<Object> getOnePedido(@PathVariable Long id) {
         try {
             return ResponseEntity.status(HttpStatus.OK).body(pedidoService.getOnePedido(id));
         } catch (RuntimeException e) {
@@ -50,7 +50,7 @@ public class PedidoController {
     }
     
     @PutMapping("/update/status/{id}")
-    public ResponseEntity<String> updateStatusPedido(@PathVariable Integer id) {
+    public ResponseEntity<String> updateStatusPedido(@PathVariable Long id) {
         try {
             pedidoService.updateStatusPedido(id);
             return ResponseEntity.status(HttpStatus.OK).body("Status do pedido atualizado com sucesso!");
@@ -60,7 +60,7 @@ public class PedidoController {
     }
 
     @PutMapping("/update/status/problema/{id}")
-    public ResponseEntity<String> updateProblemaPedido(@PathVariable Integer id) {
+    public ResponseEntity<String> updateProblemaPedido(@PathVariable Long id) {
         try {
             pedidoService.updateProblemaPedido(id);
             return ResponseEntity.status(HttpStatus.OK).body("Status do pedido atualizado para PROBLEMA_ENTREGA com sucesso!");
@@ -70,7 +70,7 @@ public class PedidoController {
     }
     
     @PutMapping("/update/status/devolucao/{id}")
-    public ResponseEntity<String> updateDevolucaoPedido(@PathVariable Integer id) {
+    public ResponseEntity<String> updateDevolucaoPedido(@PathVariable Long id) {
         try {
             Pedido pedidoAtualizado = pedidoService.updateDevolucaoPedido(id);
             return ResponseEntity.status(HttpStatus.OK).body("Status do pedido atualizado para DEVOLUCAO com sucesso!");
@@ -80,7 +80,7 @@ public class PedidoController {
     }
 
     @PutMapping("/update/status/cancelar/{id}")
-    public ResponseEntity<String> updateCancelarPedido(@PathVariable Integer id) {
+    public ResponseEntity<String> updateCancelarPedido(@PathVariable Long id) {
         try {
             pedidoService.updateCancelarPedido(id);
             return ResponseEntity.status(HttpStatus.OK).body("Pedido cancelado com sucesso!");
