+++ src/main/java/com/api/stock/repository/PedidoRepository.java
@@ -3,5 +3,5 @@ package com.api.stock.repository;
 import com.api.stock.model.Pedido;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PedidoRepository extends JpaRepository<Pedido, Integer> {
+public interface PedidoRepository extends JpaRepository<Pedido, Long> {
 }
